requirements.txt
selenium
webdriver-manager
google-cloud-storage
pymongo
python-dotenv